Fixed C++ compile errors

+++ libs/drivers/eps/eps.cpp
@@ -18,7 +18,7 @@ static I2CBus* i2c;
 
 static bool epsControlLCL(EpsLcl lcl, uint8_t state)
 {
-    uint8_t data[] = {1 + lcl, state};
+    uint8_t data[] = {static_cast<uint8_t>(1 + lcl), state};
     const I2CResult result = i2c->Write(i2c, EPS_ADDRESS, data, COUNT_OF(data));
 
     if (result != I2CResultOK)
